axvm

Trait AxVMHal

pub trait AxVMHal: Sized {
    type PagingHandler: PagingHandler;

    // Required methods
    fn alloc_memory_region_at(base: HostPhysAddr, size: usize) -> bool;
    fn dealloc_memory_region_at(base: HostPhysAddr, size: usize);
    fn virt_to_phys(vaddr: HostVirtAddr) -> HostPhysAddr;
    fn current_time_nanos() -> u64;
    fn current_vm_id() -> usize;
    fn current_vcpu_id() -> usize;
    fn current_pcpu_id() -> usize;
    fn vcpu_resides_on(vm_id: usize, vcpu_id: usize) -> AxResult<usize>;
    fn inject_irq_to_vcpu(vm_id: usize, vcpu_id: usize, irq: usize) -> AxResult;
}

The interfaces which the underlying software (kernel or hypervisor) must implement.

Required Associated Types

type PagingHandler: PagingHandler

The low-level OS-dependent helpers that must be provided for physical address management.

Required Methods

fn alloc_memory_region_at(base: HostPhysAddr, size: usize) -> bool

Allocates a memory region at the specified physical address.

Returns true if the memory region is successfully allocated.

fn dealloc_memory_region_at(base: HostPhysAddr, size: usize)

Deallocates a memory region at the specified physical address.

fn virt_to_phys(vaddr: HostVirtAddr) -> HostPhysAddr

Converts a virtual address to the corresponding physical address.

fn current_time_nanos() -> u64

Current time in nanoseconds.

fn current_vm_id() -> usize

Current VM ID.

fn current_vcpu_id() -> usize

Current Virtual CPU ID.

fn current_pcpu_id() -> usize

Current Physical CPU ID.

fn vcpu_resides_on(vm_id: usize, vcpu_id: usize) -> AxResult<usize>

Get the Physical CPU ID where the specified VCPU of the current VM resides.

Returns an error if the VCPU is not found.

fn inject_irq_to_vcpu(vm_id: usize, vcpu_id: usize, irq: usize) -> AxResult

Inject an IRQ to the specified VCPU.

This method should find the physical CPU where the specified VCPU resides and inject the IRQ to it on that physical CPU with [axvcpu::AxVCpu::inject_interrupt].

Returns an error if the VCPU is not found.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.