axvm
Module config

The configuration structure for the VM. The AxVMCrateConfig is generated from toml file, and then converted to AxVMConfig for the VM creation.

Structs

AxVCpuConfig

A part of AxVCpuConfig, which represents an architecture-dependent VCpu.

AxVMConfig

A part of AxVMCrateConfig, which represents a VM.

AxVMCrateConfig

The configuration structure for the guest VM serialized from a toml file provided by user, and then converted to AxVMConfig for the VM creation.

EmulatedDeviceConfig

A part of AxVMConfig, which represents the configuration of an emulated device for a virtual machine.

PassThroughDeviceConfig

A part of AxVMConfig, which represents the configuration of a pass-through device for a virtual machine.

VMImageConfig

A part of AxVMConfig, which stores configuration attributes related to the load address of VM images.

VmMemConfig

A part of AxVMConfig, which represents a memory region.

Enums

VMType

A part of AxVMConfig, which represents guest VM type.

VmMemMappingType

The type of memory mapping.