axvm
Trait AxVMHal

pub trait AxVMHal: Sized {
    type PagingHandler: PagingHandler;

    // Required methods
    fn alloc_memory_region_at(base: HostPhysAddr, size: usize) -> bool;
    fn dealloc_memory_region_at(base: HostPhysAddr, size: usize);
    fn virt_to_phys(vaddr: HostVirtAddr) -> HostPhysAddr;
    fn current_time_nanos() -> u64;
}

The interfaces which the underlying software (kernel or hypervisor) must implement.

Required Associated Types

type PagingHandler: PagingHandler

The low-level OS-dependent helpers that must be provided for physical address management.

Required Methods

fn alloc_memory_region_at(base: HostPhysAddr, size: usize) -> bool

Allocates a memory region at the specified physical address.

Returns true if the memory region is successfully allocated.

fn dealloc_memory_region_at(base: HostPhysAddr, size: usize)

Deallocates a memory region at the specified physical address.

fn virt_to_phys(vaddr: HostVirtAddr) -> HostPhysAddr

Converts a virtual address to the corresponding physical address.

fn current_time_nanos() -> u64

Current time in nanoseconds.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.